pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.0.1"
description = "A pytest plugin for easily instantiating reproducible mock resources."
authors = ["Prateek Pisat <pisatprateek12@gmail.com>",]
repository = "https://github.com/"
packages = [
    { include = "portfolio", from = "src" },
]
readme = 'README.md'
include = ["*.md"]

[tool.poetry.dependencies]
python = ">=3.10, <4"
pydantic = "^1.8.2"
pendulum = "^2.1.2"
Flask = "^2.0.2"
configly = "^0.2.2"
black = ">=22.3.0"
flask-pydantic-spec = "^0.4.1"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
alembic = "^1.8.1"
PyJWT = "^2.6.0"
Faker = "^15.3.4"

[tool.poetry.dev-dependencies]
black = {version = "=>19.3b0", allow-prereleases = true}
flake8 = "*"
isort = ">=5.0"
mypy = {version = "0.910", python = ">=3.5"}
pydocstyle = {version = "*", python = ">=3.5"}
pytest-mock-resources = {version = "^2.6.3", extras = ["postgres"]}


[tool.isort]
profile = 'black'
known_first_party = 'app,tests'
line_length = 100
float_to_top=true
order_by_type = false
use_parentheses = true

[tool.black]
line_length = 100

[build-system]
requires = ["poetry_core==1.0.4"]
build-backend = "poetry.core.masonry.api"
